Prototype a transport with sdmmc support
The mouthful BlockingSdioTransport generalizes an SD/MMC peripheral.
Implementations submit commands and perform data I/O. It exposes methods
to configure data lines and operation modes, and it can perform a
hardware reset with a card.

Atop of that, we build BlockingSdioHost. This dispatches commands to the
transport to initialize cards for I/O. It then exposes I/O for blocks
size of 512 bytes, concluding with an implementation of embedded-sdmmc's
BlockDevice. I came upon this approach by studing the stm32 HAL's SDMMC
implementation, along with another reference implementation for a
higher-level SD driver.

An implementation of BlockingSdioTransport is in another crate,
targeting the uSDHC peripheral available on i.MX RT processors. With
that transport implementation and this new work, I could allocate a
FAT32 volume and write a file from an iMXRT1170EVK. I'm hopeful that
this general design would also work for the stm32 HAL SDMMC driver, but
I haven't tested anything there.

This is all a proof-of-concept, and there's likely bugs. Some known
limitations, just off the cuff:

- No obvious transport interface for multiple blocks. The transport
  doesn't know how many blocks are involved in data I/O. (Is a 1024
  slice of bytes one block of 1024 bytes, or two blocks of 512 bytes?)
- I take liberties in the R2 response serialization, since my hardware
  doesn't expose the CRC and end bit to me. This means that anyone
  checking the CRC exposed through CID are going to be confused.
- The state machine is limited. It requres a CMD8 response, says that
  all voltages are OK, doesn't care for 1.8V, and generally expects an
  XC / HC card. And there's generally no consideration for (e)MMC.
- Barely any error checking for command responses.
- Took a non-optional dep on embedded-sdmmc because it was easy.