Include CRC, command index availability in Resp
In order for my uSDHC peripheral on my iMXRT hardware to perform
protocol checks, I need to know if a response should include a CRC, and
if it should include the command index. This commit extends `Resp` to
include this information.

This is a breaking change, since implementations outside this package
will need to set these flags. I opted for this approach, since it forced
me to evaluate all impls and add the flags. Another approach is to use
default values, which would be backwards compatible but may not be
correct for a hardware implementation.